extract testing utility
